my-portfolio/backend/requirements.txt
fastapi
uvicorn
twilio
gspread
python-dotenv
requests